+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - [🚀] 增加`Rect.scale`扩展;
 - [🚀] 增加`SliverStack`;
 - [🚀] `DashedLine`增加`strokeCap`参数;
+- [🚀] `SpringContainer`增加防抖(`debounceDuration`)参数;
 
 ## 0.51.2
 - [🚀] `FutureX`增加`thenIf`和`thenIfNot`扩展;