+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [🚀] 增加`SliverStack`;
 - [🚀] `DashedLine`增加`strokeCap`参数;
 - [🚀] `SpringContainer`增加防抖(`debounceDuration`)参数;
+- [🚀] 增加`computeTextLineCount`计算文本需要的行数;
 
 ## 0.51.2
 - [🚀] `FutureX`增加`thenIf`和`thenIfNot`扩展;